[Add] 검색 시 글자에 맞는 필터링 debounce 통해서 결과 도출, animate 추가

+++ KeepIt-iOS/Source/Model/ProductModel.swift
@@ -10,7 +10,7 @@ import Foundation
 struct ProductModel {
     let productImage: Data
     let productName: String
-    let productPrice: Int32
+    let productPrice: Int64
     let productLink: String
     let productMemo: String
     let productRatingStar: Double
